edit settings for multisig
